Ajout de AUTHORS, GPL, et LICENSE_BLOCK [Qu'il faudrait insérer au début de chaque fichier une fois qu'on sera d'accord sur son contenu]
Fix de deux warnings (passing float là ou attendu sf::Int8 ou int)

Ajout d'un About.

+++ PlayerShip.cpp
@@ -390,7 +390,7 @@ void PlayerShip::KonamiCodeOn()
 	overheated_ = false;
 	panel_.SetInfo("");
 	heat_ = 0.f;
-	panel_.SetHeat(heat_);	
+	panel_.SetHeat((int) heat_);	
 	coolers_ = 42;
 	panel_.SetCoolers(42);
 	hp_ = 42;
